warning修正

+++ BlockBreak/SRC/Menu.cpp
@@ -41,6 +41,6 @@ void Menu_Draw() {
 	SetFontSize(50);
 	DrawString((mgr.SCREEN_WIDTH / 2) - (50 * 3), mgr.SCREEN_HEIGHT / 2 - 150, "-BlockBreak-", 0xffffff);
 	DrawString((mgr.SCREEN_WIDTH / 2) - (50 * 4), mgr.SCREEN_HEIGHT / 2, "- PUSH [SPACE] -", 0xffffff);
-	DrawString((mgr.SCREEN_WIDTH / 2) - (50 * 4.5), mgr.SCREEN_HEIGHT / 2 + 150, "- PUSH [ESC] END -", 0xffffff);
+	DrawString((mgr.SCREEN_WIDTH / 2) - (int)(50 * 4.5), (mgr.SCREEN_HEIGHT / 2) + 150, "- PUSH [ESC] END -", 0xffffff);
 
 }
